+++ packages/discovery-components-react/.storybook/webpack.config.js
@@ -2,6 +2,9 @@ const path = require('path');
 const SRC_PATH = path.join(__dirname, '../src');
 
 module.exports = ({ config }) => {
+  config.node = {
+    fs: 'empty'
+  };
   config.module.rules.push({
     test: /\.(ts|tsx)$/,
     include: [SRC_PATH],