+++ packages/discovery-components-react/src/__tests__/components/ExampleComponent/index.test.ts
@@ -1,7 +1,0 @@
-import ExampleComponent from '../../../components/ExampleComponent/ExampleComponent';
-
-describe('ExampleComponent', () => {
-  it('is truthy', () => {
-    expect(ExampleComponent).toBeTruthy();
-  });
-});