+++ examples/discovery-components-react-example/CHANGELOG.md
@@ -3,7 +3,9 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# 1.4.0 (2019-09-11)
+## [1.9.2](https://github.ibm.com/Watson-Discovery/disco-widgets/compare/discovery-components-react-example@1.9.1...discovery-components-react-example@1.9.2) (2019-09-18)
+
+<<<<<<< HEAD
 
 ### Features
 
@@ -11,4 +13,11 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 - add e2e testing framework ([#47](https://github.ibm.com/Watson-Discovery/disco-widgets/issues/47)) ([d04eb98](https://github.ibm.com/Watson-Discovery/disco-widgets/commit/d04eb98))
 - add express server ([c71b306](https://github.ibm.com/Watson-Discovery/disco-widgets/commit/c71b306))
 - add storybook and stories for example component ([#45](https://github.ibm.com/Watson-Discovery/disco-widgets/issues/45)) ([61d02ee](https://github.ibm.com/Watson-Discovery/disco-widgets/commit/61d02ee))
-- linter ([#48](https://github.ibm.com/Watson-Discovery/disco-widgets/issues/48)) ([efe65af](https://github.ibm.com/Watson-Discovery/disco-widgets/commit/efe65af))
+- # linter ([#48](https://github.ibm.com/Watson-Discovery/disco-widgets/issues/48)) ([efe65af](https://github.ibm.com/Watson-Discovery/disco-widgets/commit/efe65af))
+  **Note:** Version bump only for package discovery-components-react-example
+
+## [1.9.1](https://github.ibm.com/Watson-Discovery/disco-widgets/compare/discovery-components-react-example@1.9.0...discovery-components-react-example@1.9.1) (2019-09-17)
+
+**Note:** Version bump only for package discovery-components-react-example
+
+> > > > > > > master