+++ packages/discovery-components-react/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.8](http:///discovery-components-react/compare/@disco-widgets/react-components@1.13.7...@disco-widgets/react-components@1.13.8) (2019-09-18)
+
+**Note:** Version bump only for package @disco-widgets/react-components
+
+## [1.13.7](http:///discovery-components-react/compare/@disco-widgets/react-components@1.13.6...@disco-widgets/react-components@1.13.7) (2019-09-17)
+
+**Note:** Version bump only for package @disco-widgets/react-components
+
+## [1.13.6](http:///discovery-components-react/compare/@disco-widgets/react-components@1.13.5...@disco-widgets/react-components@1.13.6) (2019-09-16)
+
+**Note:** Version bump only for package @disco-widgets/react-components
+
+## [1.13.5](http:///discovery-components-react/compare/@disco-widgets/react-components@1.13.4...@disco-widgets/react-components@1.13.5) (2019-09-12)
+
+**Note:** Version bump only for package @disco-widgets/react-components
+
+## [1.13.4](http:///discovery-components-react/compare/@disco-widgets/react-components@1.13.3...@disco-widgets/react-components@1.13.4) (2019-09-12)
+
+### Bug Fixes
+
+- Update @disco-widgets/react-components package readme ([#91](http:///discovery-components-react/issues/91)) ([5d3dce5](http:///discovery-components-react/commits/5d3dce5))
+
+## [1.13.3](http:///discovery-components-react/compare/@disco-widgets/react-components@1.13.1...@disco-widgets/react-components@1.13.3) (2019-09-12)
+
+**Note:** Version bump only for package @disco-widgets/react-components
+
+## [1.13.1](http:///discovery-components-react/compare/@disco-widgets/react-components@1.13.0...@disco-widgets/react-components@1.13.1) (2019-09-12)
+
+**Note:** Version bump only for package @disco-widgets/react-components
+
 # 1.13.0 (2019-09-11)
 
 ### Features